vulcan-retirejs/main.go: avoid too long file names
The check downloads all the javascript dependencies (remote and inline)
to files, which then are scanned by `retire`. Due to the way the file
name was crafted by the check, sometimes it was too long and an error
was raised.

Example:
```
https://REDACTED/ open temp/polyfill.min.js?features=Array.from%2CArray.prototype.find%2CArray.prototype.findIndex%2Ces2015%2CArray.prototype.keys%2CArray.prototype.includes%2CArray.prototype.flat%2CArray.prototype.fill%2CObject.assign%2CObject.entries%2CObject.keys%2CObject.values%2CString.prototype.endsWith%2CString.prototype.trim%2CString.prototype.padEnd%2CString.prototype.padStart%2CURLSearchParams%2CSymbol: file name too long
```

As the file name is not important for `retire` to scan it, it has been
simplified.