vulcan-trivy.go: add ScanImageMetadata option
By default Trivy does not scan the container image metadata, it scans
just the filesystem. Secrets might be present in the metadata too, for
example when injected as `ARG` parameters in the Dockerfile followed by
`RUN` commands.

This commit adds a new option to enable scanning the container image
metadata for secrets, as specified in
https://aquasecurity.github.io/trivy/v0.45/docs/target/container_image/