vulcan-trivy.go: increase trivy CLI timeout
The trivy CLI has a default 5 minutes timeout, that was extended to 10
minutes for the case of git repositories. The problem is that executions
of the trivy command timing out might finish as `FAILED` check status
while it should be `TIMEOUT`.

In order to change that behaviour, all trivy commands should run with a
timeout value bigger than the timeout value defined in the check
manifest. In that way the check will always timeout before the command
does, and therefore will make the check status be `TIMEOUT`, making
easier to analyze check statistics.