Update addons on build time and disable in runtime
When updating addons the server opens port 8080 before finishing updating
the addons. In fact the logs shows two times the message about ZAP listening
on port 8080 twice.
The check starts creating the scan before ZAP is ready and fails.
By disabling with -silent we can start creating the scan just when the port
is open.